requirements.txt
fastapi
uvicorn[standard]
jinja2
psycopg2-binary